Creating adapters' folder
