+++ app/src/main/java/com/example/ecodigify/ui/adapters/IngredientFragmentListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.ecodigify
+package com.example.ecodigify.ui.adapters
 
 import android.content.res.ColorStateList
 import android.view.LayoutInflater
@@ -8,6 +8,7 @@ import android.widget.ProgressBar
 import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
+import com.example.ecodigify.R
 import com.example.ecodigify.dataclass.Ingredient
 import java.time.LocalDate
 import java.time.Period
