+++ app/src/main/java/com/example/ecodigify/ui/adapters/RecipeFragmentListAdapter.kt
@@ -1,36 +1,35 @@
-package com.example.ecodigify
+package com.example.ecodigify.ui.adapters
 
-import android.content.res.ColorStateList
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.ecodigify.dataclass.RecipeFull
+import com.example.ecodigify.R
+import com.example.ecodigify.dataclass.Recipe
 
-class RecipeFullFragmentListAdapter(private val dataSet: Array<RecipeFull>, private val onClick: (RecipeFull) -> Unit) :
-    RecyclerView.Adapter<RecipeFullFragmentListAdapter.ViewHolder>() {
+class RecipeFragmentListAdapter(private val dataSet: Array<Recipe>, private val onClick: (Recipe) -> Unit) :
+    RecyclerView.Adapter<RecipeFragmentListAdapter.ViewHolder>() {
 
     /**
      * Provide a reference to the type of views that you are using
      * (custom ViewHolder)
      */
-    class ViewHolder(view: View, val onClick: (RecipeFull) -> Unit) : RecyclerView.ViewHolder(view) {
+    class ViewHolder(view: View, val onClick: (Recipe) -> Unit) : RecyclerView.ViewHolder(view) {
         // Define click listener for the ViewHolder's View
-        val imageView: ImageView = view.findViewById(R.id.recipeFullImageView)
-        val textView: TextView = view.findViewById(R.id.recipeFullTitleTextView)
-        val recyclerView: RecyclerView = view.findViewById(R.id.recipeFullIngredientsRecyclerView)
+        val imageView: ImageView = view.findViewById(R.id.recipeImageView)
+        val textView: TextView = view.findViewById(R.id.recipeTitleTextView)
 
-        private var currentRecipeFull: RecipeFull? = null
+        private var currentRecipe: Recipe? = null
 
         init {
-            itemView.setOnClickListener { currentRecipeFull?.let { recipeFull -> onClick(recipeFull) } }
+            itemView.setOnClickListener { currentRecipe?.let { recipe -> onClick(recipe) } }
         }
 
-        fun bind(recipeFull: RecipeFull) {
-            currentRecipeFull = recipeFull
-            textView.text = recipeFull.name
+        fun bind(recipe: Recipe) {
+            currentRecipe = recipe
+            textView.text = recipe.name
         }
     }
 
@@ -52,7 +51,6 @@ class RecipeFullFragmentListAdapter(private val dataSet: Array<RecipeFull>, priv
 
         // viewHolder.imageView // TODO: implement image retrival
         viewHolder.textView.text = dataSet[position].name
-        // TODO: populate recycler view
     }
 
     // Return the size of your dataset (invoked by the layout manager)
