+++ app/src/main/java/com/example/ecodigify/ui/adapters/RecipeFullFragmentListAdapter.kt
@@ -1,40 +1,36 @@
-package com.example.ecodigify
+package com.example.ecodigify.ui.adapters
 
-import android.content.res.ColorStateList
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
-import android.widget.ProgressBar
 import android.widget.TextView
-import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import com.example.ecodigify.dataclass.Recipe
-import java.time.LocalDate
-import java.time.Period
-import kotlin.math.min
+import com.example.ecodigify.R
+import com.example.ecodigify.dataclass.RecipeFull
 
-class RecipeFragmentListAdapter(private val dataSet: Array<Recipe>, private val onClick: (Recipe) -> Unit) :
-    RecyclerView.Adapter<RecipeFragmentListAdapter.ViewHolder>() {
+class RecipeFullFragmentListAdapter(private val dataSet: Array<RecipeFull>, private val onClick: (RecipeFull) -> Unit) :
+    RecyclerView.Adapter<RecipeFullFragmentListAdapter.ViewHolder>() {
 
     /**
      * Provide a reference to the type of views that you are using
      * (custom ViewHolder)
      */
-    class ViewHolder(view: View, val onClick: (Recipe) -> Unit) : RecyclerView.ViewHolder(view) {
+    class ViewHolder(view: View, val onClick: (RecipeFull) -> Unit) : RecyclerView.ViewHolder(view) {
         // Define click listener for the ViewHolder's View
-        val imageView: ImageView = view.findViewById(R.id.recipeImageView)
-        val textView: TextView = view.findViewById(R.id.recipeTitleTextView)
+        val imageView: ImageView = view.findViewById(R.id.recipeFullImageView)
+        val textView: TextView = view.findViewById(R.id.recipeFullTitleTextView)
+        val recyclerView: RecyclerView = view.findViewById(R.id.recipeFullIngredientsRecyclerView)
 
-        private var currentRecipe: Recipe? = null
+        private var currentRecipeFull: RecipeFull? = null
 
         init {
-            itemView.setOnClickListener { currentRecipe?.let { recipe -> onClick(recipe) } }
+            itemView.setOnClickListener { currentRecipeFull?.let { recipeFull -> onClick(recipeFull) } }
         }
 
-        fun bind(recipe: Recipe) {
-            currentRecipe = recipe
-            textView.text = recipe.name
+        fun bind(recipeFull: RecipeFull) {
+            currentRecipeFull = recipeFull
+            textView.text = recipeFull.name
         }
     }
 
@@ -56,6 +52,7 @@ class RecipeFragmentListAdapter(private val dataSet: Array<Recipe>, private val 
 
         // viewHolder.imageView // TODO: implement image retrival
         viewHolder.textView.text = dataSet[position].name
+        // TODO: populate recycler view
     }
 
     // Return the size of your dataset (invoked by the layout manager)
